requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
dateparser==1.0.0
requests==2.26.0